+++ src/gg.regression.unity.bots/Runtime/Scripts/ClientDashboard/TcpMessage.cs.meta
